requirements.txt
python-telegram-bot
python-dotenv
urllib3
youtube-dl